exos_facts refactor
