+++ lib/ansible/module_utils/network/exos/facts/facts.py
@@ -12,9 +12,14 @@ calls the appropriate facts gathering function
 from ansible.module_utils.network.exos.argspec.facts.facts import FactsArgs
 from ansible.module_utils.network.common.facts.facts import FactsBase
 from ansible.module_utils.network.exos.facts.lldp_global.lldp_global import Lldp_globalFacts
+from ansible.module_utils.network.exos.facts.legacy.base import Default, Hardware, Interfaces, Config
 
+FACT_LEGACY_SUBSETS = dict(
+    default=Default,
+    hardware=Hardware,
+    interfaces=Interfaces,
+    config=Config)
 
-FACT_LEGACY_SUBSETS = {}
 FACT_RESOURCE_SUBSETS = dict(
     lldp_global=Lldp_globalFacts,
 )
